refactor(webapp): extract shared bespoke Modal shell
