+++ packages/webapp/src/components/board/ExitDialog.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { createPortal } from 'react-dom';
 import { useRouter } from 'next/navigation';
-import { StickerButton } from '@/components/design';
+import { Modal, StickerButton } from '@/components/design';
 import { getLobbyErrorMessage, leaveRoom } from '@/app/lobby/actions';
 import { clearCredentials, loadCredentials } from '@/lib/matchCredentials';
 
@@ -10,10 +9,7 @@ import { clearCredentials, loadCredentials } from '@/lib/matchCredentials';
  * 回大廳 keeps the seat + credentials so the lobby offers 回到桌子 (#421);
  * 離開座位 releases the seat via leaveMatch, which terminates the match for
  * everyone (remaining players see the terminated overlay); 取消 does nothing.
- *
- * Bespoke .modal overlay (no MUI): portaled to <body> to escape sticky-header
- * stacking contexts, with Escape/backdrop close, initial focus, a small Tab
- * cycle, and body scroll lock while open.
+ * Rendered in the shared bespoke Modal shell (no MUI).
  */
 export default function ExitDialog({
   open,
@@ -27,46 +23,7 @@ export default function ExitDialog({
   const router = useRouter();
   const [isLeaving, setIsLeaving] = React.useState(false);
   const [error, setError] = React.useState<string | null>(null);
-  const panelRef = React.useRef<HTMLDivElement>(null);
   const canClose = !isLeaving;
-
-  React.useEffect(() => {
-    if (!open) {
-      return;
-    }
-    panelRef.current?.focus();
-    const previousOverflow = document.body.style.overflow;
-    document.body.style.overflow = 'hidden';
-    return () => {
-      document.body.style.overflow = previousOverflow;
-    };
-  }, [open]);
-
-  const handleKeyDown = (event: React.KeyboardEvent) => {
-    if (event.key === 'Escape') {
-      if (canClose) {
-        onClose();
-      }
-      return;
-    }
-    if (event.key !== 'Tab') {
-      return;
-    }
-    // Keep Tab cycling inside the three buttons.
-    const focusables = panelRef.current?.querySelectorAll<HTMLElement>('button:not(:disabled)');
-    if (!focusables || focusables.length === 0) {
-      return;
-    }
-    const first = focusables[0];
-    const last = focusables[focusables.length - 1];
-    if (event.shiftKey && (document.activeElement === first || document.activeElement === panelRef.current)) {
-      event.preventDefault();
-      last.focus();
-    } else if (!event.shiftKey && document.activeElement === last) {
-      event.preventDefault();
-      first.focus();
-    }
-  };
 
   const handleLobbyKeepSeat = () => {
     router.push('/lobby');
@@ -91,71 +48,59 @@ export default function ExitDialog({
     }
   };
 
-  if (!open) {
-    return null;
-  }
-
-  return createPortal(
-    <div className="modal-backdrop" onClick={canClose ? onClose : undefined}>
-      <div
-        ref={panelRef}
-        className="modal"
-        role="dialog"
-        aria-modal="true"
-        aria-label="離開遊戲"
-        tabIndex={-1}
-        data-testid="exit-dialog"
-        style={{ width: 'min(420px, 100%)', fontFamily: 'var(--font-zh)', outline: 'none' }}
-        onClick={(event) => event.stopPropagation()}
-        onKeyDown={handleKeyDown}
-      >
-        <div style={{ display: 'flex', alignItems: 'center', gap: 10, marginBottom: 6 }}>
-          <span aria-hidden style={{ fontSize: 22 }}>🚪</span>
-          <div>
-            <strong style={{ fontSize: 16, color: 'var(--ink)' }}>要離開嗎？</strong>
-            <div className="tag-en">Leave the game?</div>
-          </div>
-        </div>
-        <p style={{ fontSize: 13, color: 'var(--ink-soft)', lineHeight: 1.5, margin: '6px 0 16px' }}>
-          你可以回到大廳並保留座位，隨時回來繼續；或直接離席（這局遊戲會終止）。
-        </p>
-        {error && (
-          <p role="alert" style={{ fontSize: 12, fontWeight: 700, color: 'var(--orange-deep)', margin: '0 0 12px' }}>
-            ⚠ {error}
-          </p>
-        )}
-        <div style={{ display: 'flex', flexDirection: 'column', gap: 10 }}>
-          <StickerButton
-            variant="teal"
-            onClick={handleLobbyKeepSeat}
-            disabled={isLeaving}
-            style={{ width: '100%' }}
-            data-testid="exit-keep-seat"
-          >
-            回大廳（保留座位） <span className="tag-en" style={{ color: 'rgba(255,255,255,0.85)' }}>Go to lobby</span>
-          </StickerButton>
-          <StickerButton
-            variant="ghost"
-            onClick={() => void handleLeaveSeat()}
-            disabled={isLeaving}
-            style={{ width: '100%', borderColor: 'var(--orange-deep)', color: 'var(--orange-deep)' }}
-            data-testid="exit-leave-seat"
-          >
-            {isLeaving ? '離開中…' : '離開座位（終止遊戲）'} <span className="tag-en">Leave seat</span>
-          </StickerButton>
-          <StickerButton
-            variant="ghost"
-            size="sm"
-            onClick={onClose}
-            disabled={isLeaving}
-            style={{ width: '100%' }}
-            data-testid="exit-cancel"
-          >
-            取消 <span className="tag-en">Cancel</span>
-          </StickerButton>
+  return (
+    <Modal
+      open={open}
+      onClose={canClose ? onClose : undefined}
+      ariaLabel="離開遊戲"
+      width="min(420px, 100%)"
+      dataTestid="exit-dialog"
+    >
+      <div style={{ display: 'flex', alignItems: 'center', gap: 10, marginBottom: 6 }}>
+        <span aria-hidden style={{ fontSize: 22 }}>🚪</span>
+        <div>
+          <strong style={{ fontSize: 16, color: 'var(--ink)' }}>要離開嗎？</strong>
+          <div className="tag-en">Leave the game?</div>
         </div>
       </div>
-    </div>,
-    document.body,
+      <p style={{ fontSize: 13, color: 'var(--ink-soft)', lineHeight: 1.5, margin: '6px 0 16px' }}>
+        你可以回到大廳並保留座位，隨時回來繼續；或直接離席（這局遊戲會終止）。
+      </p>
+      {error && (
+        <p role="alert" style={{ fontSize: 12, fontWeight: 700, color: 'var(--orange-deep)', margin: '0 0 12px' }}>
+          ⚠ {error}
+        </p>
+      )}
+      <div style={{ display: 'flex', flexDirection: 'column', gap: 10 }}>
+        <StickerButton
+          variant="teal"
+          onClick={handleLobbyKeepSeat}
+          disabled={isLeaving}
+          style={{ width: '100%' }}
+          data-testid="exit-keep-seat"
+        >
+          回大廳（保留座位） <span className="tag-en" style={{ color: 'rgba(255,255,255,0.85)' }}>Go to lobby</span>
+        </StickerButton>
+        <StickerButton
+          variant="ghost"
+          onClick={() => void handleLeaveSeat()}
+          disabled={isLeaving}
+          style={{ width: '100%', borderColor: 'var(--orange-deep)', color: 'var(--orange-deep)' }}
+          data-testid="exit-leave-seat"
+        >
+          {isLeaving ? '離開中…' : '離開座位（終止遊戲）'} <span className="tag-en">Leave seat</span>
+        </StickerButton>
+        <StickerButton
+          variant="ghost"
+          size="sm"
+          onClick={onClose}
+          disabled={isLeaving}
+          style={{ width: '100%' }}
+          data-testid="exit-cancel"
+        >
+          取消 <span className="tag-en">Cancel</span>
+        </StickerButton>
+      </div>
+    </Modal>
   );
 }
