+++ packages/webapp/src/components/design/index.ts
@@ -1,6 +1,7 @@
 export { default as AppHeader } from './AppHeader';
 export { default as Logo } from './Logo';
 export { default as StickerButton } from './StickerButton';
+export { default as Modal } from './Modal';
 export { default as PaperCard } from './PaperCard';
 export { default as CharacterAvatar } from './CharacterAvatar';
 export {
